+++ packages/vui-wysiwyg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vui-wysiwyg
 
+## 6.0.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.18.24
+
 ## 6.0.23
 
 ### Patch Changes